Implement event-triggered flows
Refactor the integration logic, to use application events as
input. Add a feed fetching job to demonstrate this capability.

Attempt to reduce coupling by using an interface for messages,
so they can be identified and converted to canonical formats.

Also improve on the publisher-subscriber logic, so no duplicate
payloads can reach the targets.