Implement stand-alone Twitch stream announcer
Improve the implementation of stream announcements, refactoring it
to a new entity, with its own service and resource classes, and its
own view.

Disable unneeded check for duplicate message and refactored the
previous poller job into a new one that calls the added service.

Modified the event object to delegate most of the logic done to the
service class.

Fix a couple of issues with how the job processed settings, such as
the one used to control how many minutes have to pass to announce a
previously announced stream.

Add usability features to improve ease of use when adding streams.